# [IMP] Compacting the creation of Journal Lines for CABA base lines.