Main
-
- Configure Tax based on payment.
- Have one invoice with one hundred lines
- Pay the invoice, be it partial or in full, it doesn't matter.
- Check the CABA entry created:
Chances are that the Journal Entry will end up having:
-- 1 Entry line per tax collected.
-- 1 Entry line per tax to be collected.
-- 200 Entry lines for the base.

Odoo does not economize on resources when creating CABA Entries.

This PR deals with the creation of a more Slimmed way of creating the CABA Entries.
After this PR you will end up with 

-- 1 Entry line per tax collected.
-- 1 Entry line per tax to be collected.
-- 2 Entry lines for the base, one debit, one credit.

chances are that this could be increased but the algorithm tries only to group by a set of keys.

Regards.

YouTube
-

https://youtu.be/njXJDJLOFWc

Zoom
-

Grabación despues de aplicar el fix.
https://zoom.us/rec/share/ypBlI4zcqERJTJXisU_webUsJqr0X6a81XJI_PRczhufQPE0zor1bM7fWrNzCDlo



--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr