@hbto 2020-02-02
@moylop260 

This PR Tries to address issue 7881

This will not fix the issue but in the long term will avoid it.

The problem is not the algorithm in https://git.vauxoo.com/vauxoo/mexico/blob/85753e311bccd9bd8b39613e89cce05b907bb85f/l10n_mx_avoid_reversal_entry/models/account_move.py#L48

The problem is odoo:

Let us have 4 invoices with 200 lines each one.
Do a payment.
This creates a CABA Entry with  4 x (200 x 2 + 2) = 808 Caba lines.

Unreconcile the Payment.
It has now to delete 808 Caba lines plus all the checks needed for deletions.

cc/ @nhomar