@moylop260 2020-02-02
FYI I'm using this branch:

1) Cancel the lodi payment 11819
2) Set to draft
3) Confirm

Result:
```
  File "/home/odoo/enterprise/l10n_mx_edi/models/account_payment.py", line 96, in post
    l10n_mx_edi_manual_reconciliation=False)).post()
  File "/home/odoo/odoo-11.0/addons/account/models/account_payment.py", line 531, in post
    move = rec._create_payment_entry(amount)
  File "/home/odoo/odoo-11.0/addons/account/models/account_payment.py", line 604, in _create_payment_entry
    self.invoice_ids.register_payment(counterpart_aml)
  File "/home/odoo/odoo-11.0/addons/account/models/account_invoice.py", line 941, in register_payment
    return (line_to_reconcile + payment_line).reconcile(writeoff_acc_id, writeoff_journal_id)
  File "/home/odoo/enterprise/l10n_mx_edi/models/account_move.py", line 12, in reconcile
    writeoff_acc_id, writeoff_journal_id)
  File "/home/odoo/odoo-11.0/addons/account/models/account_move.py", line 1092, in reconcile
    remaining_moves = self.auto_reconcile_lines()
  File "/home/odoo/odoo-11.0/addons/account/models/account_move.py", line 1062, in auto_reconcile_lines
    'currency_id': currency,
  File "/home/odoo/odoo-11.0/addons/account/models/account_move.py", line 2014, in create
    res.create_tax_cash_basis_entry(percentage_before_rec)
  File "/.repo_requirements/virtualenv/python3.6/lib/python3.6/site-packages/line_profiler/line_profiler.py", line 122, in wrapper
    result = func(*args, **kwds)
  File "/home/odoo/odoo-11.0/addons/account/models/account_move.py", line 1898, in create_tax_cash_basis_entry
    self._create_tax_cash_basis_base_line(caba_keys, caba_amount_dict, caba_amount_currency_dict, newly_created_move)
  File "/home/odoo/odoo-11.0/addons/account/models/account_move.py", line 1803, in _create_tax_cash_basis_base_line
    base_line = self._get_tax_cash_basis_base_line(key)
TypeError: _get_tax_cash_basis_base_line() missing 1 required positional argument: 'new_move'
```

So I just applied the following patch:
```patch
diff --git a/addons/account/models/account_move.py b/addons/account/models/account_move.py
index 66f5ffeef..16480ed8d 100644
--- a/addons/account/models/account_move.py
+++ b/addons/account/models/account_move.py
@@ -1233,7 +1233,7 @@ class AccountMoveLine(models.Model):
             'exchange_move_id': exchange_move.id if exchange_move else False,
         })

-    @api.multi
+    @profile
     def remove_move_reconcile(self):
         """ Undo a reconciliation """
         if not self:
@@ -1800,7 +1800,7 @@ class AccountPartialReconcile(models.Model):

     def _create_tax_cash_basis_base_line(self, keys, amount_dict, amount_currency_dict, new_move):
         for key in keys:
-            base_line = self._get_tax_cash_basis_base_line(key)
+            base_line = self._get_tax_cash_basis_base_line(key, new_move)
             rounded_amt = amount_dict[key]
             amount_currency = amount_currency_dict[key]
             aml_obj = self.env['account.move.line'].with_context(check_move_validity=False)
@@ -1816,6 +1816,7 @@ class AccountPartialReconcile(models.Model):
                 amount_currency=-amount_currency,
                 tax_ids=[]))

+    @profile
     def create_tax_cash_basis_entry(self, percentage_before_rec):
         self.ensure_one()
         move_date = self.debit_move_id.date
```